utils/requirements.txt
requests==2.19.1
gitpython==2.1.11
fire==0.1.3
six==1.11.0
